requirements_lite.txt
openai
PyMuPDF
Pillow
numpy
requests
gradio
python-dotenv
email-validator>=2.0
tqdm
